requirements.txt
pandas>=2.0
openpyxl>=3.1
python-dateutil>=2.9
phonenumbers>=8.13
